feat: Server validations for CRUD operations

+++ restapi/src/controllers/places/PlacesController.ts
@@ -1,54 +1,69 @@
 import {Request, Response} from "express";
 import {PlaceType} from "../../types/PlaceType";
 import Place from "../../models/Place";
+import {validationResult} from "express-validator";
 
-const getPlaces = async (req: Request, res: Response): Promise<void> => {
+const getPlaces = async (req: Request, res: Response): Promise<Response> => {
     try {
         const places: Array<PlaceType> = await Place.find();
-        res.send(places);
+        return res.status(200).send(places);
     }
     catch (error) {
-        console.log("An error has occurred while retrieving the list of places: " + error)
+        return res.status(500).send("An error has occurred while retrieving the list of places: \n\n" + error)
     }
 }
 
-const addPlace = async (req: Request, res: Response): Promise<void> => {
+const addPlace = async (req: Request, res: Response): Promise<Response> => {
     try {
+        const errors = validationResult(req);
+        if (!errors.isEmpty()) {
+            return res.status(400).json({errors: errors.array()})
+        }
         await Place.create({
             title: req.body.title,
             uuid: req.body.uuid,
             longitude: req.body.longitude,
             latitude: req.body.latitude
         });
-        res.sendStatus(200)
+        return res.sendStatus(200)
     }
     catch (error) {
-        console.log("An error has occurred while adding a place with title " + req.params.title + ": " + error)
+        return res.status(500).send("An error has occurred while adding a place: " + error)
     }
 }
 
-const deletePlace = async (req: Request, res: Response): Promise<void> => {
+const deletePlace = async (req: Request, res: Response): Promise<Response> => {
     try {
+        const errors = validationResult(req);
+        if (!errors.isEmpty()) {
+            return res.status(400).json({errors: errors.array()})
+        }
         await Place.deleteOne({title: req.params.title})
-        res.sendStatus(200)
+        return res.status(200)
     }
     catch (error) {
-        console.log("An error has occurred while deleting a place with title " + req.params.title + ": " + error)
+        return res.status(500).send("An error has occurred while deleting a place with title " +
+            req.params.title + ": " + error)
     }
 }
 
-const updatePlace = async (req: Request, res: Response): Promise<void> => {
+const updatePlace = async (req: Request, res: Response): Promise<Response> => {
     try {
+        const errors = validationResult(req);
+        if (!errors.isEmpty()) {
+            return res.status(400).json({errors: errors.array()})
+        }
         await Place.findOneAndUpdate({title: req.params.title}, {
             title: req.body.title,
             uuid: req.body.uuid,
             longitude: req.body.longitude,
             latitude: req.body.latitude
         });
-        res.sendStatus(200)
+        return res.status(200)
     }
     catch (error) {
-        console.log("An error has occurred while updating a place with title " + req.params.title + ": " + error)
+        return res.status(500).send("An error has occurred while updating a place with title " +
+            req.params.title + ": " + error)
     }
 }
 export { getPlaces, addPlace, deletePlace, updatePlace }
